requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.7.14
charset-normalizer==3.4.2
chromedriver-binary==74.0.3729.6.0
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
Django==5.2.4
djangorestframework==3.16.0
exceptiongroup==1.3.0
flake8==7.3.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
murmurhash==1.0.13
mypy_extensions==1.1.0
mysqlclient==2.2.7
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
preshed==3.0.10
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
selenium==4.34.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
textblob==0.19.0
thinc==8.3.6
tomli==2.2.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
tweepy==4.16.0
tweety==0.1.6
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vaderSentiment==3.3.2
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
